ver.0/requirements.txt
# Core Framework & Database
Flask==2.3.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# LlamaIndex RAG System
llama-index==0.8.36
llama-index-embeddings-openai==0.1.3
llama-index-llms-openai==0.1.3
llama-index-vector-stores-chroma==0.1.3

# AI & Machine Learning
openai==0.28.1
chromadb==0.4.13
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0

# Vector Database & Search
pgvector==0.2.3
faiss-cpu==1.7.4

# Web Framework Extensions
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Data Processing & Analysis
pandas==2.0.3
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3

# HTTP & API
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5

# Caching & Performance
redis==4.6.0
celery==5.3.1

# Data Validation & Serialization
marshmallow==3.20.1
jsonschema==4.19.0

# Logging & Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
PyYAML==6.0.1

# Testing (Development)
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0
factory-boy==3.3.0
faker==19.6.2

# Code Quality (Development)
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1

# Security
cryptography==41.0.4
bcrypt==4.0.1
PyJWT==2.8.0

# Documentation (Optional)
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Production Server
gunicorn==21.2.0
uvicorn==0.23.2
